Persist Album and Genre Data

+++ utils.rb
@@ -19,14 +19,20 @@ OPTIONS = {
 def load_data
   recover_author
   recover_games
+  load_genres
+  load_albums
 end
 
 def preserve_data
   save_author(Author.all)
   save_game(Game.all)
+  save_genres
+  save_albums
 end
 
 def recover_author
+  return unless File.exist?('data/authors.json') && File.size?('data/authors.json')
+
   File.write('./data/authors.json', '') unless File.exist?('./data/authors.json')
   author_data = File.read('./data/authors.json')
 
@@ -45,6 +51,8 @@ def recover_objects_by_id(id)
 end
 
 def recover_games
+  return unless File.exist?('data/games.json') && File.size?('data/games.json')
+
   File.write('./data/games.json', '') unless File.exist?('./data/games.json')
   games_data = File.read('./data/games.json')
 
@@ -61,4 +69,49 @@ def recover_games
       new_game.multiplayer = game[3]
     end
   end
+end
+
+def load_genres
+  return unless File.exist?('data/genres.json') && File.size?('data/genres.json')
+
+  JSON.parse(File.read('data/genres.json')).each { |genre| Genre.new(genre['name'], genre['id']) }
+end
+
+def load_albums
+  return unless File.exist?('data/albums.json') && File.size?('data/albums.json')
+
+  JSON.parse(File.read('data/albums.json')).each do |album|
+    genre_obj = Genre.all.find { |genre| genre.id == album['genre_id'] }
+    author_obj = Author.all.find { |author| author.id == album['author_id'] }
+    MusicAlbum.new(genre_obj, author_obj, album['label'], album['publish_date'], on_spotify: album['on_spotify'])
+  end
+end
+
+def save_genres
+  File.exist?('data/genres.json') ? File.open('data/genres.json', 'w') : File.new('data/genres.json', 'w')
+  File.write('data/genres.json', Genre.all.to_json)
+end
+
+def save_albums
+  File.exist?('data/albums.json') ? File.open('data/albums.json', 'w') : File.new('data/albums.json', 'w')
+  File.write('data/albums.json', MusicAlbum.all.to_json)
+end
+
+def album_from_user_input
+  print 'Input album author first name: '
+  author_firstname = gets.chomp
+  print 'Input album author last name: '
+  author_lastname = gets.chomp
+  author = Author.all.find { |aut| aut.first_name == author_firstname && aut.last_name == author_lastname }
+  author = Author.new(author_firstname, author_lastname) if author.nil?
+  print 'Input album label: '
+  label = gets.chomp
+  print 'Input album genre: '
+  genre_name = gets.chomp
+  genre = Genre.all.find { |gen| gen.name == genre_name }
+  genre = Genre.new(genre_name) if genre.nil?
+  print 'Input publish date: '
+  publish_date = gets.chomp
+  MusicAlbum.new(genre, author, label, publish_date)
+  puts "Album created succesfully.\n\n"
 end
